fixing height of tables
