+++ packages/vue/src/components/BlockCsrTable/CsrAttachment.vue
